Move HttpException and its subclasses to common HTTP module
