+++ archetypes/helidon/src/main/archetype/se/custom/files/src/main/java/__pkg__/FileStorage.java.multipart.mustache
@@ -1,8 +1,8 @@
 
 package {{package}};
 
-import io.helidon.reactive.webserver.BadRequestException;
-import io.helidon.reactive.webserver.NotFoundException;
+import io.helidon.common.http.BadRequestException;
+import io.helidon.common.http.NotFoundException;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
