+++ common/http/src/main/java/io/helidon/common/http/HttpException.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.reactive.webserver;
-
-import io.helidon.common.http.Http;
+package io.helidon.common.http;
 
 /**
  * Runtime exception for applications.
