+++ examples/translator-app/backend/src/main/java/io/helidon/examples/translator/backend/TranslatorBackendService.java
@@ -18,7 +18,7 @@ package io.helidon.examples.translator.backend;
 import java.util.HashMap;
 import java.util.Map;
 
-import io.helidon.reactive.webserver.BadRequestException;
+import io.helidon.common.http.BadRequestException;
 import io.helidon.reactive.webserver.Routing;
 import io.helidon.reactive.webserver.ServerRequest;
 import io.helidon.reactive.webserver.ServerResponse;
