+++ reactive/webserver/jersey/src/main/java/io/helidon/reactive/webserver/jersey/JerseySupport.java
@@ -37,10 +37,10 @@ import io.helidon.common.configurable.ThreadPool;
 import io.helidon.common.context.Context;
 import io.helidon.common.context.Contexts;
 import io.helidon.common.http.Http;
+import io.helidon.common.http.HttpException;
 import io.helidon.config.Config;
 import io.helidon.config.ConfigValue;
 import io.helidon.reactive.webserver.Handler;
-import io.helidon.reactive.webserver.HttpException;
 import io.helidon.reactive.webserver.KeyPerformanceIndicatorSupport;
 import io.helidon.reactive.webserver.Routing;
 import io.helidon.reactive.webserver.ServerRequest;
