+++ reactive/webserver/webserver/src/main/java/io/helidon/reactive/webserver/RequestRouting.java
@@ -35,7 +35,9 @@ import io.helidon.common.LazyValue;
 import io.helidon.common.context.Contexts;
 import io.helidon.common.http.HtmlEncoder;
 import io.helidon.common.http.Http;
+import io.helidon.common.http.HttpException;
 import io.helidon.common.http.HttpMediaType;
+import io.helidon.common.http.NotFoundException;
 import io.helidon.tracing.Span;
 import io.helidon.tracing.SpanContext;
 import io.helidon.tracing.Tracer;
